% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direstraits_palette.R
\name{direstraits_palette}
\alias{direstraits_palette}
\title{Dire Straits Palette}
\usage{
direstraits_palette(palette = "money_for_nothing", ...)
}
\arguments{
\item{palette}{Select a color palette from dire straits album covers}

\item{...}{More arguments}
}
\value{
direstraits_palette
}
\description{
Create a color palette from Dire Straits album covers
}
