% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direstraits_palette_gen.R
\name{direstraits_palette_gen}
\alias{direstraits_palette_gen}
\title{A helper function generated for the Dire Straits Palette}
\usage{
direstraits_palette_gen(palette = "money_for_nothing", direction = 1)
}
\arguments{
\item{palette}{The color palette from Dire Straits album covers.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors are order as appeared in the palette.
If -1, the order of colors is reversed.}
}
\value{
direstraits_palette_gen
}
\description{
A helper function that is used in the color scales.
}
