% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_palette.R
\name{mix_palette}
\alias{mix_palette}
\title{Discrete color palettes from a variety of music bands}
\usage{
mix_palette(palette = "radio_remba", ...)
}
\arguments{
\item{palette}{The color palette from mixed album covers.

   The list of available palettes includes: "estacion_mexico" (manu chao), "radio_remba" (manu chao; default),
   "auberge" (Chris Rea), "best_cris_rea1994" (Chris Rea), "best_cris_rea2009" (Chris Rea),
   "one_fine_day" (Chris Rea), "saudade" (Thievery Corporation), "richest_man_in_babylon" (Thievery Corporation),
   "treasures" (Thievery Corporation), "industrial_silence" (madrugada), "grit" (madrugada), "inevitable_end" (royksopp).}

\item{...}{Other arguments passed on to the function.}
}
\value{
mix_palette
}
\description{
It includes color palettes from a number of different album covers.
   It allows us to select one of the available palettes. The function returns a vector
   with the hex codes of the selected palette.
}
