% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_palette_gen.R
\name{mix_palette_gen}
\alias{mix_palette_gen}
\title{A helper function generated from a variety of music bands}
\usage{
mix_palette_gen(palette = "radio_remba", direction = 1)
}
\arguments{
\item{palette}{The color palette from mixed album covers.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors are order as appeared in the palette.
If -1, the order of colors is reversed.}
}
\value{
mix_palette_gen
}
\description{
A helper function that is used in the color scales for mix palettes.
}
