% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinkfloyd_palette.R
\name{pinkfloyd_palette}
\alias{pinkfloyd_palette}
\title{Discrete color palettes from Pink Floyd music band.}
\usage{
pinkfloyd_palette(palette = "wish_you_were_here", ...)
}
\arguments{
\item{palette}{The color palette from Pink Floyd album covers.

   The list of available palettes includes: "echoes", "the_final_cut", "division_bell",
   "animals", "wish_you_were_here" (default), "momentary_lapse_Of_reason", "meddle", "atom_heart_mother",
   "more", "endless_river".}

\item{...}{Other arguments passed on to the function.}
}
\value{
pinkfloyd_palette
}
\description{
It includes color palettes from Pink Floyd album covers.
   It allows us to select one of the available palettes. The function returns a vector
   with the hex codes of the selected palette.
}
