% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinkfloyd_palette_gen.R
\name{pinkfloyd_palette_gen}
\alias{pinkfloyd_palette_gen}
\title{A helper function generated for the Pink Floyd palettes}
\usage{
pinkfloyd_palette_gen(palette = "wish_you_were_here", direction = 1)
}
\arguments{
\item{palette}{The color palette from Pink Floyd album covers.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors are order as appeared in the palette.
If -1, the order of colors is reversed.}
}
\value{
pinkfloyd_palette_gen
}
\description{
A helper function that is used in the color scales for Pink Floyd palettes.
}
