% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_direstraits.R
\name{scale_color_direstraits}
\alias{scale_color_direstraits}
\alias{scale_colour_direstraits}
\alias{scale_fill_direstraits}
\title{Color scales functions for the Dire Straits}
\usage{
scale_color_direstraits(palette = "money_for_nothing", direction = 1, ...)

scale_colour_direstraits(palette = "money_for_nothing", direction = 1, ...)

scale_fill_direstraits(palette = "money_for_nothing", direction = 1, ...)
}
\arguments{
\item{palette}{The color palette from Dire Straits album covers.

   The list of available palettes includes: "on_the_night", "so_far_away"
   "brothers_in_arms", "communique", "money_for_nothing" (default), "dire_straits",
   "on_every_street", "alchemy", "studio_99", "making_movies", "love_over_gold",
   "roterdam", and "very_best".}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors are order as appeared in the palette.
If -1, the order of colors is reversed.}

\item{...}{Other arguments passed on to the color scale functions.}
}
\value{
scale_color_direstraits

scale_colour_direstraits

scale_fill_direstraits
}
\description{
Color scales functions for the Dire Straits palettes.
}
\details{
color palette from dire straits

colour palette from dire straits

fill palette from dire straits
}
