% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_mix.R
\name{scale_color_mix}
\alias{scale_color_mix}
\alias{scale_colour_mix}
\alias{scale_fill_mix}
\title{Color scales functions from a variety of music bands}
\usage{
scale_color_mix(palette = "radio_remba", direction = 1, ...)

scale_colour_mix(palette = "radio_remba", direction = 1, ...)

scale_fill_mix(palette = "radio_remba", direction = 1, ...)
}
\arguments{
\item{palette}{The color palette from mixed album covers.

   The list of available palettes includes: "estacion_mexico" (manu chao), "radio_remba" (manu chao; default),
   "auberge" (Chris Rea), "best_cris_rea1994" (Chris Rea), "best_cris_rea2009" (Chris Rea),
   "one_fine_day" (Chris Rea), "saudade" (Thievery Corporation), "richest_man_in_babylon" (Thievery Corporation),
   "treasures" (Thievery Corporation), "industrial_silence" (madrugada), "grit" (madrugada), "inevitable_end" (royksopp).}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors are order as appeared in the palette.
If -1, the order of colors is reversed.}

\item{...}{Other arguments passed on to the color scale functions.}
}
\value{
scale_color_mix

scale_colour_mix

scale_fill_mix
}
\description{
Discrete color scales functions for the mix palettes.
}
\details{
color palette from mixed albums

colour palette from mixed albums

fill palette from mixed albums
}
