% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_pinkfloyd.R
\name{scale_color_pinkfloyd}
\alias{scale_color_pinkfloyd}
\alias{scale_colour_pinkfloyd}
\alias{scale_fill_pinkfloyd}
\title{Color scales functions for the Pink Floyd}
\usage{
scale_color_pinkfloyd(palette = "wish_you_were_here", direction = 1, ...)

scale_colour_pinkfloyd(palette = "wish_you_were_here", direction = 1, ...)

scale_fill_pinkfloyd(palette = "wish_you_were_here", direction = 1, ...)
}
\arguments{
\item{palette}{The color palette from Pink Floyd album covers.

   The list of available palettes includes: "echoes", "the_final_cut", "division_bell",
   "animals", "wish_you_were_here" (default), "momentary_lapse_Of_reason",
   "meddle", "atom_heart_mother", "more", "endless_river".}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors are order as appeared in the palette.
If -1, the order of colors is reversed.}

\item{...}{Other arguments passed on to the color scale functions.}
}
\value{
scale_color_pinkfloyd

scale_colour_pinkfloyd

scale_fill_pinkfloyd
}
\description{
Discrete color scales functions for the Pink Floyd palettes.
}
\details{
color palette from pink floyd

colour palette from pink floyd

fill palette from pink floyd
}
