% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direstraits_palette.R
\name{direstraits_palette}
\alias{direstraits_palette}
\title{Dire Straits Palette}
\usage{
direstraits_palette(palette = "money_for_nothing", ...)
}
\arguments{
\item{palette}{The name of color palette from Dire Straits album covers.

   The list of available palettes includes: "on_the_night", "so_far_away"
   "brothers_in_arms", "communique", "money_for_nothing" (default), "dire_straits",
   "on_every_street", "alchemy", "studio_99", "making_movies", "love_over_gold",
   "roterdam", and "very_best".}

\item{...}{Other arguments passed on to the function.}
}
\value{
direstraits_palette
}
\description{
It includes color palettes generated from the Dire Straits album covers.
   It allows us to select one of the available palettes. The function returns a vector
   with the hex codes of the selected palette.
}
