% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direstraits_palette_gen.R
\name{direstraits_palette_gen}
\alias{direstraits_palette_gen}
\title{A helper function generated for the Dire Straits Palette}
\usage{
direstraits_palette_gen(palette = "money_for_nothing", direction = 1)
}
\arguments{
\item{palette}{Select a color palette from Dire Straits album covers.}

\item{direction}{Select the direction of the color order.}
}
\value{
direstraits_palette_gen
}
\description{
A helper function that will be used in the color scales.
}
